+++ tests/testthat/test-PomaBoxplots.R
@@ -64,6 +64,7 @@ test_that("PomaBoxplots works", {
   expect_error(PomaBoxplots(norm_ls, group = "features", feature_name = "hello"))
   expect_error(PomaBoxplots(norm_ls, group = "features", feature_name = "methyl_succina"))
   expect_error(PomaBoxplots(norm_ls, group = "features", feature_name = c("methyl_succina", "linoleic_acid")))
+  expect_error(PomaBoxplots(norm_none, group = "features", jitter = FALSE, label_size = 15, legend_position = "no"))
   
 })